Reindent files
