+++ lib/domain/models/scope-editor.model.ts
@@ -1,7 +1,7 @@
 export interface ScopeEditor {
-    id: string;
-    scopeId: string;
-    userId?: string | null;
-    email?: string | null;
-    createdAt: Date;
+  id: string;
+  scopeId: string;
+  userId?: string | null;
+  email?: string | null;
+  createdAt: Date;
 }
