+++ lib/domain/models/scope.model.ts
@@ -1,13 +1,13 @@
 export interface Scope {
-    id: string;
-    name: string;
-    description?: string;
-    ownerId: string;
-    createdAt: Date;
-    updatedAt: Date;
+  id: string;
+  name: string;
+  description?: string;
+  ownerId: string;
+  createdAt: Date;
+  updatedAt: Date;
 }
 
 export enum ScopeType {
-    OWNED = 'owned',
-    SHARED = 'shared'
+  OWNED = 'owned',
+  SHARED = 'shared'
 }
