+++ lib/domain/models/team-member.model.ts
@@ -1,0 +1,8 @@
+export interface TeamMember {
+  id: string;
+  scopeId: string;
+  name: string;
+  role: string;
+  fte: number;
+  createdAt: Date;
+}
