--- lib/domain/models/team_member.model.ts
@@ -1,8 +1,0 @@
-export interface TeamMember {
-    id: string;
-    scopeId: string;
-    name: string;
-    role: string;
-    fte: number;
-    createdAt: Date;
-}
