+++ lib/domain/models/user.model.ts
@@ -1,13 +1,13 @@
 export interface User {
-    id: string;
-    email: string;
-    fullName?: string;
-    avatarUrl?: string;
-    createdAt: Date;
-    updatedAt: Date;
-    emailConfirmedAt?: Date | null;
-    invitedAt?: Date | null;
+  id: string;
+  email: string;
+  fullName?: string;
+  avatarUrl?: string;
+  createdAt: Date;
+  updatedAt: Date;
+  emailConfirmedAt?: Date | null;
+  invitedAt?: Date | null;
 
-    // eslint-disable-next-line
-    additional: Record<string, any>;
+  // eslint-disable-next-line
+  additional: Record<string, any>;
 }
