+++ lib/domain/repositories/user.repository.ts
@@ -1,5 +1,5 @@
-import {User} from "@/lib/domain/models/user.model";
+import { User } from "@/lib/domain/models/user.model";
 
 export abstract class UserRepository {
-    abstract getLoggedInUser(): Promise<User | null>;
+  abstract getLoggedInUser(): Promise<User | null>;
 }
