+++ lib/domain/services/remove-scope.service.ts
@@ -1,23 +1,23 @@
-import {ScopeEditorRepository} from "@/lib/domain/repositories/scope-editor.repository";
-import {injectable} from "inversify";
-import {User} from "@/lib/domain/models/user.model";
+import { ScopeEditorRepository } from "@/lib/domain/repositories/scope-editor.repository";
+import { injectable } from "inversify";
+import { User } from "@/lib/domain/models/user.model";
 
 @injectable()
 export class RemoveScopeService {
 
-    private readonly scopeEditorRepository: ScopeEditorRepository;
+  private readonly scopeEditorRepository: ScopeEditorRepository;
 
-    constructor(
-        scopeEditorRepository: ScopeEditorRepository,
-    ) {
-        this.scopeEditorRepository = scopeEditorRepository;
+  constructor(
+    scopeEditorRepository: ScopeEditorRepository,
+  ) {
+    this.scopeEditorRepository = scopeEditorRepository;
+  }
+
+  public async removeScope(user: User, scopeId: string): Promise<void> {
+    if (user?.email === null) {
+      return
     }
 
-    public async removeScope(user: User, scopeId: string): Promise<void> {
-        if (user?.email === null) {
-            return
-        }
-
-        await this.scopeEditorRepository.deleteByScopeId({scopeId: scopeId, userId: user.id});
-    }
+    await this.scopeEditorRepository.deleteByScopeId({ scopeId: scopeId, userId: user.id });
+  }
 }
